requirements.txt
yfinance==0.2.55
pandas==1.5.0
numpy==1.23.4
matplotlib==3.6.0
seaborn==0.12.0
requests==2.32.3
fpdf==1.7.2
schedule==1.1.0
feedparser==6.0.11
python-dotenv==1.1.0